+++ packages/cli/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Zowe CLI package will be documented in this file.
 
 ## Recent Changes
 
--Enhancement: Added new command zowe zos-files download all-members-matching, (zowe files dl amm), to download members matching specified pattern(s). [#2175](https://github.com/zowe/zowe-cli/issues/2175)
+-Enhancement: Added new command zowe zos-files download all-members-matching, (zowe files dl amm), to download members matching specified pattern(s). [#2359](https://github.com/zowe/zowe-cli/pull/2359)
 
 ## Recent Changes